requirements.txt
psutil
adafruit-circuitpython-ssd1306
pillow
board
digitalio

